requirements.txt
-e .  # executes setup.py -> 'src' becomes a (local) package

boto3==1.16.52
ipython-sql==0.4.0
jupyter==1.0.0
psycopg2==2.8.6
python-dotenv==0.15.0
